requirements.txt
requests
python-dotenv
tzdata
python-dateutil
packaging
